pyproject.toml
[project]
name = "worduelist-bot"
version = "0.0.1"
description = "A bot for solving codes for Samuel Britnell's Worduel: https://gamejolt.com/games/worduel/685260"
authors = [{ name = "Elijah Sink", email = "theminorprofit@gmail.com" }]
dependencies = ["python-telegram-bot>=13.11", "requests>=2.27.1"]
requires-python = ">=3.10"
license = { text = "MIT" }

[project.urls]
homepage = ""

[tool]
[tool.pdm]
[tool.pdm.scripts]
_.env_file = ".env"
dev = "nodemon main.py"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
